chore: Extracted option-parser
