+++ src/index.js
@@ -1,6 +1,7 @@
 import { render } from 'react-dom'
 import { navigate } from './api'
 import App from './app'
+import OptionsParser from './api/options'
 
 function init(element, options) {
   if (typeof element === 'string') {
@@ -10,33 +11,11 @@ function init(element, options) {
   render(<App {...options} />, element)
 }
 
-function getOptions(element) {
-  const options = {}
-
-  Array.prototype.slice.call(element.attributes)
-    .filter(attr => attr.name.startsWith('data-'))
-    .map(attr => { options[getName(attr)] = getValue(attr) })
-
-  return options
-}
-
-function getName(attribute) {
-  return attribute.name
-    .slice(5)
-    .replace(/-([a-z])/g, match => {
-      return match[1].toUpperCase()
-    })
-}
-
-function getValue(attribute) {
-  return attribute.value == '' ? true : attribute.value
-}
-
 document.addEventListener('DOMContentLoaded', () => {
   const element = document.querySelector('[data-endpoint]')
 
   if (element) {
-    init(element, getOptions(element))
+    init(element, OptionsParser(element))
 
     const url = element.getAttribute('data-endpoint')
